Remove color export (#3868)
