+++ src/components/ExceptionList/ExceptionList.tsx
@@ -1,8 +1,7 @@
 import React, {Fragment} from 'react';
 
 import {classNames, variationName} from '../../utilities/css';
-import {Icon} from '../Icon';
-import type {IconProps} from '../../types';
+import {Icon, IconProps} from '../Icon';
 import {Truncate} from '../Truncate';
 
 import styles from './ExceptionList.scss';
