+++ src/components/Modal/components/Dialog/Dialog.tsx
@@ -4,13 +4,13 @@ import {Transition, CSSTransition} from 'react-transition-group';
 
 import {classNames} from '../../../../utilities/css';
 import {focusFirstFocusableNode} from '../../../../utilities/focus';
-import {AnimationProps, Key} from '../../../../types';
+import {Key} from '../../../../types';
 import {KeypressListener} from '../../../KeypressListener';
 import {TrapFocus} from '../../../TrapFocus';
 
 import styles from './Dialog.scss';
 
-interface BaseDialogProps {
+export interface DialogProps {
   labelledBy?: string;
   instant?: boolean;
   children?: React.ReactNode;
@@ -19,9 +19,8 @@ interface BaseDialogProps {
   onClose(): void;
   onEntered?(): void;
   onExited?(): void;
+  in?: boolean;
 }
-
-export type DialogProps = BaseDialogProps & AnimationProps;
 
 export function Dialog({
   instant,
