+++ src/components/Navigation/components/Section/Section.tsx
@@ -6,8 +6,7 @@ import {navigationBarCollapsed} from '../../../../utilities/breakpoints';
 import {useUniqueId} from '../../../../utilities/unique-id';
 import {useToggle} from '../../../../utilities/use-toggle';
 import {Collapsible} from '../../../Collapsible';
-import {Icon} from '../../../Icon';
-import type {IconProps} from '../../../../types';
+import {Icon, IconProps} from '../../../Icon';
 import {Item, ItemProps} from '../Item';
 import styles from '../../Navigation.scss';
