+++ src/components/OptionList/OptionList.tsx
@@ -3,7 +3,7 @@ import React, {useState, useCallback} from 'react';
 import {arraysAreEqual} from '../../utilities/arrays';
 import {useFeatures} from '../../utilities/features';
 import {classNames} from '../../utilities/css';
-import type {IconProps} from '../../types';
+import type {IconProps} from '../Icon';
 import type {AvatarProps} from '../Avatar';
 import type {ThumbnailProps} from '../Thumbnail';
 import {useUniqueId} from '../../utilities/unique-id';
